created animations
I created multiple different animations for different states/actions including idle, and jumping. The jumping one took me awhile because I decided to make the character rotate 360 degrees when I jump, so I had to create the sprite for each frame, hence ~70 frames, with 5 degree rotation shift for each. The idle animation is a more simple one,  where the character just shifts a little bit. The "run" animation is actually a lack of animation, because I needed to add a "run" state to make the transitions work correctly from state to state.